Apply SELinux labels to the persist filesystem.
Presently, the persist filesystem remains
unlabeled when mounted. This patch introduces
a series of restorecons in the init.mako.rc file
to properly label the persist filesystem using
a small number of new labels.

Note: Since the persist filesystem is ext4 it
is possible to label these files at build
time and then remove the numerous restorecons.

Change-Id: I48eaa2b9901ac8c978192c14493ba1058a089423
Signed-off-by: rpcraig <rpcraig@tycho.ncsc.mil>