Fix /data/misc/audio pathname regex.
It is a regex, not a glob, so use (/.*)? rather than /* to
match the directory and anything beneath it.

Also, in our policy, we had this as a generic audio_data_file and in
the core sepolicy rather than mako-specific.  Consider moving this
to external/sepolicy and possibly renaming the type to be more general
(although this will require a restorecon_recursive).

Change-Id: Ib8c96ab9e19d34e8e34a4c859528345763be4906
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>