Increase headset mic volume on Nexus 4
The volume of the microphone when using headset is too low on the Nexus 4 as reported on several forums :
 - http://forum.xda-developers.com/showthread.php?t=2513603
 - https://code.google.com/p/android/issues/detail?id=41860
 - https://productforums.google.com/forum/#!topic/nexus/dOaR3XxCmxo

This patch only modifies the headset microphone gain so that it gets louder (volume of 99 instead of the 84 by default).
It does not change anything else.
It solves the problem on my Nexus 4.

Change-Id: I0ad10062688b49fb87c4d2271f6835abda7e6f8a
Signed-off-by: Stéphane Rosi <stephane.rosi@gmail.com>