Increase headset mic volume on Nexus 4
The volume of the microphone when using headset is too low on the Nexus 4 as reported on several forums :
 - http://forum.xda-developers.com/showthread.php?t=2513603
 - https://code.google.com/p/android/issues/detail?id=41860
 - https://productforums.google.com/forum/#!topic/nexus/dOaR3XxCmxo

This patch only modifies the headset microphone gain so that it gets louder.
It replicates the patch used for Nexus 5 : https://android.googlesource.com/device/lge/hammerhead/+/76d97e731c73cad9e5f46501fdc752fe3cb9b31a%5E%21/#F0
It does not change anything else.
It solves the problem on my Nexus 4.

--
06/01/14 : Patch modified as per comment
07/01/2014 : Patch modified to replicate Nexus 5 patch.

Change-Id: I2d01e66cc3fcb92cad1be1caba0e7ba13d82d064